replaced code with 13933 from Ramesh

+++ src/components/AppComponents/TimeoutPopup/timeOutUtils.tsx
@@ -1,5 +1,5 @@
-import { logout } from '@pega/react-sdk-components/lib/components/helpers/authManager';
-import { getSdkConfig } from '@pega/react-sdk-components/lib/components/helpers/config_access';
+import { getSdkConfig } from '@pega/auth/lib/sdk-auth-manager';
+import { triggerLogout } from '../../helpers/utils';
 
 let milisecondsTilWarning = 780 * 1000;
 let milisecondsTilSignout = 115 * 1000;
@@ -20,10 +20,15 @@ export function clearTimer() {
   clearTimeout(signoutTimeout);
 }
 
-export const initTimeout = (showTimeoutModal, deleteData, isAuthorised) => {
+export const initTimeout = async (
+  showTimeoutModal,
+  deleteData,
+  isAuthorised,
+  isConfirmationPage
+) => {
   // TODO - isAuthorised to be replaced by caseType from pega
   // Fetches timeout length config
-  settingTimer();
+  await settingTimer();
   clearTimeout(applicationTimeout);
   clearTimeout(signoutTimeout);
 
@@ -32,12 +37,34 @@ export const initTimeout = (showTimeoutModal, deleteData, isAuthorised) => {
     // TODO - unauth and sessiontimeout functionality to be implemented
     showTimeoutModal(true);
     signoutTimeout = setTimeout(() => {
-      if (isAuthorised) {
-        logout();
-      } else {
+      if (!isAuthorised && !isConfirmationPage && deleteData) {
+        // if the journey is not authorized or from confirmation page , the claim data gets deleted
         deleteData();
         clearTimer();
         // session ends and deleteData() (pega)
+      }
+    }, milisecondsTilSignout);
+  }, milisecondsTilWarning);
+};
+
+export const resetTimeout = (showTimeoutModal, deleteData, isAuthorised, isConfirmationPage) => {
+  // TODO - isAuthorised to be replaced by caseType from pega
+  // Fetches timeout length config
+  clearTimeout(applicationTimeout);
+  clearTimeout(signoutTimeout);
+
+  // Clears any existing timeouts and starts the timeout for warning, after set time shows the modal and starts signout timer
+  applicationTimeout = setTimeout(() => {
+    showTimeoutModal(true);
+    signoutTimeout = setTimeout(() => {
+      if (!isAuthorised && !isConfirmationPage && deleteData) {
+        // if the journey is not authorized or from confirmation page , the claim data gets deleted
+        deleteData();
+        clearTimer();
+      } else {
+        // the logout case executes when entire timeout occurs after confirmation page or user clicks
+        // exit survey link in pop after confirmation page
+        triggerLogout();
       }
     }, milisecondsTilSignout);
   }, milisecondsTilWarning);
@@ -49,12 +76,15 @@ export function staySignedIn(
   claimsListApi,
   deleteData = null,
   isAuthorised = false,
-  refreshSignin = true
+  refreshSignin = true,
+  isConfirmationPage = false
 ) {
-  if (refreshSignin) {
+  if (refreshSignin && !!claimsListApi) {
     // @ts-ignore
-    PCore.getDataPageUtils().getDataAsync(claimsListApi, 'root');
+    PCore.getDataPageUtils().getDataAsync(claimsListApi, 'root', {
+      OperatorId: 'Model_Unauth@ChB'
+    });
   }
   setShowTimeoutModal(false);
-  initTimeout(setShowTimeoutModal, deleteData, isAuthorised);
+  resetTimeout(setShowTimeoutModal, deleteData, isAuthorised, isConfirmationPage);
 }
